move isSliceInt64Eq to util pkg; improve function names & typo
